Move all casters to customCasters.
