+++ cpp/tensorrt_llm/pybind/bindings.cpp
@@ -37,6 +37,7 @@
 #include "tensorrt_llm/pybind/runtime/bindings.h"
 #include "tensorrt_llm/pybind/userbuffers/bindings.h"
 #include "tensorrt_llm/runtime/common.h"
+#include "tensorrt_llm/runtime/cudaStream.h"
 #include "tensorrt_llm/runtime/gptJsonConfig.h"
 #include "tensorrt_llm/runtime/ipcNvlsMemory.h"
 #include "tensorrt_llm/runtime/ipcUtils.h"
@@ -101,6 +102,16 @@ PYBIND11_MODULE(TRTLLM_PYBIND_MODULE, m)
                 auto& world = tensorrt_llm::mpi::MpiComm::world();
                 tensorrt_llm::mpi::MpiComm::setSession(world.split(color, rank));
             });
+
+    py::classh<tr::CudaStream>(m, "CudaStream")
+        .def(py::init(
+                 [](py::object py_stream)
+                 {
+                     cudaStream_t stream = reinterpret_cast<cudaStream_t>(py_stream.cast<uintptr_t>());
+                     return tr::CudaStream{stream};
+                 }),
+            py::arg("stream_ptr"))
+        .def("get_device", &tr::CudaStream::getDevice);
 
     // Create submodule for executor bindings.
     py::module_ executor_submodule = m.def_submodule("executor", "Executor bindings");
