+++ cpp/tensorrt_llm/pybind/executor/executorConfig.cpp
@@ -18,17 +18,15 @@
 #include "executorConfig.h"
 #include "tensorrt_llm/executor/executor.h"
 #include "tensorrt_llm/executor/types.h"
-
+#include "tensorrt_llm/runtime/cudaStream.h"
+#include "tensorrt_llm/runtime/utils/mpiUtils.h"
+#include <optional>
 #include <pybind11/cast.h>
 #include <pybind11/functional.h>
 #include <pybind11/operators.h>
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
-
-#include "streamCaster.h"
-#include "tensorCaster.h"
-
-#include <optional>
+#include <torch/torch.h>
 #include <vector>
 
 namespace py = pybind11;
@@ -306,6 +304,7 @@ void initConfigBindings(pybind11::module_& m)
 
     auto logitsPostProcessorConfigGetstate = [](tle::LogitsPostProcessorConfig const& self)
     { return py::make_tuple(self.getProcessorMap(), self.getProcessorBatched(), self.getReplicate()); };
+
     auto logitsPostProcessorConfigSetstate = [](py::tuple const& state)
     {
         if (state.size() != 3)
